backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
email-validator
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
